Persistent storage for samba4 when configured as an AD DC
Ticket: #3901

+++ nanobsd/Files/usr/local/libexec/nas/generate_smb4_conf.py
@@ -4,6 +4,7 @@ import os
 import string
 import sys
 import tempfile
+import time
 
 sys.path.extend([
     '/usr/local/www',
@@ -562,7 +563,25 @@ def provision_smb4():
     if p.returncode != 0:
         return False
 
-    return True
+    try: 
+        pfile = "/var/db/samba4/.provisioned"
+        with open(pfile, 'w') as f:
+            f.close()
+        os.chmod(pfile, 0400)
+        return True
+
+    except Exception as e:
+        print e
+
+    return False
+
+
+def smb4_domain_provisioned():
+    pfile = "/var/db/samba4/.provisioned"
+    if os.path.exists(pfile) and os.path.isfile(pfile):
+        return True
+
+    return False
 
 
 def smb4_mkdir(dir):
@@ -580,11 +599,12 @@ def smb4_unlink(dir):
 
 
 def smb4_setup():
+    statedir = "/var/db/samba4"
+
     smb4_mkdir("/var/run/samba")
     smb4_mkdir("/var/db/samba")
 
     smb4_mkdir("/var/run/samba4")
-    smb4_mkdir("/var/db/samba4")
 
     smb4_mkdir("/var/log/samba4")
     os.chmod("/var/log/samba4", 0755)
@@ -594,6 +614,31 @@ def smb4_setup():
 
     smb4_unlink("/usr/local/etc/smb.conf")
     smb4_unlink("/usr/local/etc/smb4.conf")
+
+    role = get_server_role()
+    if role != 'dc':
+        smb4_mkdir(statedir)
+          
+    elif role == 'dc':
+        dc = DomainController.objects.all()[0]
+        if os.path.exists(statedir) and os.path.islink(statedir):
+            return
+
+        elif os.path.exists(statedir):
+            try:
+                p = pipeopen("/bin/rm -rf '%s'" % statedir)
+                p.communicate()
+
+            except:
+                olddir = "%s.%s" % (statedir, time.time())
+                p = pipeopen("/bin/mv '%s' '%s'" % (statedir, olddir))
+                p.communicate()
+
+        try:
+            os.symlink(dc.dc_storage, statedir)
+        except Exception as e:  
+            print "Unable to create symlink '%s' -> '%s' (%s)" % (
+                dc.dc_storage, statedir, e)
 
 
 def main():
@@ -610,7 +655,7 @@ def main():
     generate_smb4_shares(smb4_shares)
 
     role = get_server_role()
-    if role == 'dc':
+    if role == 'dc' and not smb4_domain_provisioned():
         provision_smb4()
          
     with open(smb_conf_path, "w") as f:
@@ -625,13 +670,14 @@ def main():
         os.write(fd, line + '\n')
     os.close(fd)
 
-    p = pipeopen("/usr/local/bin/pdbedit -d 0 -i smbpasswd:%s -e %s -s %s" % (
-        tmpfile, "tdbsam:/var/etc/private/passdb.tdb", smb_conf_path))
-    out = p.communicate()
-    if out and out[1]:
-        for line in out[1].split('\n'):
-            print line
-    os.unlink(tmpfile)
+    if role != 'dc':
+        p = pipeopen("/usr/local/bin/pdbedit -d 0 -i smbpasswd:%s -e %s -s %s" % (
+            tmpfile, "tdbsam:/var/etc/private/passdb.tdb", smb_conf_path))
+        out = p.communicate()
+        if out and out[1]:
+            for line in out[1].split('\n'):
+                print line
+        os.unlink(tmpfile)
 
 
 if __name__ == '__main__':
